jban/requirements.txt
psycopg2
django_extensions
django_rest_framework
